requirements.txt
google-genai
google-adk
streamlit